Move resolver APIs into adapter classes
