+++ gradle.properties
@@ -1,6 +1,6 @@
 basename=xmlresolver
 
-resolverVersion=6.0.0
+resolverVersion=6.0.1
 
 group=org.xmlresolver
