+++ src/main/java/org/xmlresolver/adapters/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Implementations of various third-party resolver APIs.
+ * <p>The classes in this package implement APIs for SAX, DOM, StAX, and Xerces XNI resolvers.</p>
+ * <p>They're in separate classes to avoid a compile-time dependency on unused APIs when
+ * instantiating an {@link org.xmlresolver.XMLResolver}</p>
+ */
+
+package org.xmlresolver.adapters;
